Completes Bloccit Authorization Assignment

+++ app/controllers/posts_controller.rb
@@ -1,4 +1,5 @@
 class PostsController < ApplicationController
+  include PostsHelper
 
   before_action :require_sign_in, except: :show
   before_action :authorize_user, except: [:show, :new, :create]
@@ -8,8 +9,13 @@ class PostsController < ApplicationController
   end
 
   def new
-    @topic = Topic.find(params[:topic_id])
-    @post = Post.new
+    if !user_is_moderator?
+      @topic = Topic.find(params[:topic_id])
+      @post = Post.new
+    else
+      flash[:alert] = "Only members and admins can create posts."
+      redirect_to topics_path
+    end
   end
 
   def edit
@@ -30,31 +36,41 @@ class PostsController < ApplicationController
   end
 
   def create
-    @topic = Topic.find(params[:topic_id])
-    @post = @topic.posts.build(post_params)
+    if !user_is_moderator?
+      @topic = Topic.find(params[:topic_id])
+      @post = @topic.posts.build(post_params)
 
-    @post.topic = @topic
-    @post.user = current_user
+      @post.topic = @topic
+      @post.user = current_user
 
-    if @post.save
+      if @post.save
 
-      flash[:notice] = "Post was saved"
-      redirect_to [@topic, @post]
+        flash[:notice] = "Post was saved"
+        redirect_to [@topic, @post]
+      else
+        flash.now[:alert] = "There was an error saving your post. Please try again."
+        render :new
+      end
     else
-      flash.now[:alert] = "There was an error saving your post. Please try again."
-      render :new
+      flash[:alert] = "Only members and admins can create posts."
+      redirect_to topics_path
     end
   end
 
   def destroy
-    @post = Post.find(params[:id])
+    if !user_is_moderator?
+      @post = Post.find(params[:id])
 
-    if @post.destroy
-      flash[:notice] = "\"#{@post.title}\" was deleted succesfully."
-      redirect_to @post.topic
+      if @post.destroy
+        flash[:notice] = "\"#{@post.title}\" was deleted succesfully."
+        redirect_to @post.topic
+      else
+        flash.now[:alert] = "There was an error deleting the post."
+        render :show
+      end
     else
-      flash.now[:alert] = "There was an error deleting the post."
-      render :show
+      flash[:alert] = "Only members and admins can destroy posts."
+      render :show 
     end
   end
 
@@ -65,10 +81,10 @@ class PostsController < ApplicationController
 
   def authorize_user
     post = Post.find(params[:id])
-    unless current_user == post.user || current_user.admin?
+    unless current_user == post.user || current_user.admin? || user_is_moderator?
       flash[:alert] = "You must be an admin to do that."
       redirect_to [post.topic, post]
-    end 
+    end
   end
 
 end
